+++ PalestreGoGO/DB/dbo/Tables/TipologieClienti.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[TipologieClienti]
 (
-	[Id]					SMALLINT		NOT NULL,
+	[Id]					INT				NOT NULL,
 	[Nome]					NVARCHAR(50)	NOT NULL,
 	[Descrizione]			NVARCHAR(100)	NULL,
